requirements.txt
Django==1.6.2
dj-database-url==0.2.2
dj-static==0.0.5
django-braces==1.4.0
django-crispy-forms==1.4.0
django-registration==1.0
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
six==1.8.0
static==1.0.2
wsgiref==0.1.2
yolk==0.4.3
